+++ sdk/spring/spring-cloud-azure-service/src/test/java/com/azure/spring/cloud/service/implementation/servicebus/factory/ServiceBusReceiverClientBuilderFactoryTests.java
@@ -59,6 +59,8 @@ class ServiceBusReceiverClientBuilderFactoryTests
         final ServiceBusClientBuilder.ServiceBusReceiverClientBuilder builder = factory.build();
         builder.buildClient();
 
+        verify(getSharedServiceBusClientBuilder(properties), times(1)).customEndpointAddress(customEndpoint);
+
         verify(builder, times(1)).topicName("test-topic");
         verify(builder, times(1)).subscriptionName("test-subscription");
         verify(builder, times(1)).receiveMode(ServiceBusReceiveMode.PEEK_LOCK);
@@ -70,7 +72,7 @@ class ServiceBusReceiverClientBuilderFactoryTests
         verify(factory.getServiceBusClientBuilder(), times(1)).fullyQualifiedNamespace(properties.getFullyQualifiedNamespace());
     }
 
-    private static ServiceBusReceiverClientTestProperties getServiceBusReceiverClientTestProperties() {
+    private ServiceBusReceiverClientTestProperties getServiceBusReceiverClientTestProperties() {
         ServiceBusReceiverClientTestProperties properties = new ServiceBusReceiverClientTestProperties();
         properties.setNamespace("test-namespace");
         properties.setEntityName("test-topic");
@@ -81,6 +83,7 @@ class ServiceBusReceiverClientBuilderFactoryTests
         properties.setPrefetchCount(100);
         properties.setMaxAutoLockRenewDuration(Duration.ofSeconds(5));
         properties.setAutoComplete(false);
+        properties.setCustomEndpointAddress(this.customEndpoint);
         return properties;
     }