+++ sdk/spring/spring-cloud-azure-service/src/test/java/com/azure/spring/cloud/service/implementation/servicebus/factory/TestServiceBusClientBuilderFactory.java
@@ -11,8 +11,7 @@ import static org.mockito.Mockito.mock;
 
 public class TestServiceBusClientBuilderFactory extends ServiceBusClientBuilderFactory {
     public TestServiceBusClientBuilderFactory(AzureProperties azureProperties) {
-
-        super((ServiceBusClientCommonProperties)azureProperties);
+        super((ServiceBusClientCommonProperties) azureProperties);
     }
 
     @Override